extract SceneObject createObject-related action creators from Toolbar
